pyproject.toml
[project]
name = "shoppy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.42.2",
    "django>=5.1.6",
    "django-filter>=25.1",
    "django-taggit>=6.1.0",
    "pillow>=11.1.0",
]
